- augment ListUtils with FastRemoveAt that auto decrements a ref index - augment ListUtils with a TryGetValidElement which removes an item null   or with null attributes that must not be null - update code with ref syntax for indexes for List.FastRemoveAt - update Activation for SwivelController so that swivels inside a   Vehicle will still be added as a temp piece for location/zone syncing.
