+++ src/ValheimRAFT/AssetBundle/AssetBundle.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 3321472274
+CRC: 2325733271
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
