+++ src/ValheimRAFT/AssetBundle/valheim-vehicles.manifest
@@ -1,15 +1,15 @@
 ManifestFileVersion: 0
-CRC: 2556103617
+CRC: 3694322110
 Hashes:
   AssetFileHash:
     serializedVersion: 2
-    Hash: 82f8756f760b6198bea5f7d1a84b2637
+    Hash: 488b3d502690d6eac026b5411536b8da
   TypeTreeHash:
     serializedVersion: 2
     Hash: cbca681f9021208bbcc6d59371405a02
   IncrementalBuildHash:
     serializedVersion: 2
-    Hash: 82f8756f760b6198bea5f7d1a84b2637
+    Hash: 488b3d502690d6eac026b5411536b8da
 HashAppended: 0
 ClassTypes:
 - Class: 1
